+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 -   [INF] Update pre-commit hooks and remove mutable references. Issue #844. @loganthomas
 -   [INF] Add GitHub Release pointer to auto-release script. Issue #818. @loganthomas
 -   [INF] Updated black version in github actions code-checks to match pre-commit hooks. @nvamsikrishna05
+-   [ENH] Add reset_index flag to row_to_names function. @fireddd
 -   [ENH] Updated `label_encode` to use pandas factorize instead of scikit-learn LabelEncoder. @nvamsikrishna05
 -   [INF] Removed the scikit-learn package from the dependencies from environment-dev.yml and base.in files. @nvamsikrishna05
 -   [ENH] Add function to remove constant columsn. @fireddd